Change log file location for queue worker
The location of the log file for the queue worker in the Docker configuration has been changed from a file located at /var/www/storage/logs/worker.log to standard error output /dev/stderr.
This change is to follow Docker's best practices, logs should be sent to the stdout or stderr streams to handle by the logging driver.